requirements.txt
django==3.2
celery
djangorestframework
mysql-client